+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Attribute 'subnet_of' used at dual stack network no longer only
   applies to IPv4 address. It now declares, that IPv4 oder IPv6 or
   both addresses are in subnet relation.
+- It is no longer mandatory for a virtual or secondary interface of
+  dual stack interface to be dual stack as well.
 
 ### Fixed